+++ lib/generators/rails_toastify/install/templates/rails_toastify.rb
@@ -1,0 +1,3 @@
+RailsToastify.setup do |config|
+  config.framework = :tailwind #or :bootstrap
+end